refactor: 테스트 코드에서 통합 테스트용 api를 IntegrationFixture 클래스 하위로 이동
