+++ backend/src/test/java/edonymyeon/backend/comment/integration/CommentIntegrationTest.java
@@ -1,11 +1,11 @@
 package edonymyeon.backend.comment.integration;
 
-import static edonymyeon.backend.comment.integration.steps.CommentSteps.게시물에_대한_댓글을_모두_조회한다;
-import static edonymyeon.backend.comment.integration.steps.CommentSteps.댓글을_삭제한다;
-import static edonymyeon.backend.comment.integration.steps.CommentSteps.댓글을_생성한다;
 import static edonymyeon.backend.global.exception.ExceptionInformation.COMMENT_ID_NOT_FOUND;
 import static edonymyeon.backend.global.exception.ExceptionInformation.COMMENT_MEMBER_NOT_SAME;
 import static edonymyeon.backend.global.exception.ExceptionInformation.POST_ID_NOT_FOUND;
+import static edonymyeon.backend.support.IntegrationFixture.CommentSteps.게시물에_대한_댓글을_모두_조회한다;
+import static edonymyeon.backend.support.IntegrationFixture.CommentSteps.댓글을_삭제한다;
+import static edonymyeon.backend.support.IntegrationFixture.CommentSteps.댓글을_생성한다;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
