+++ backend/src/test/java/edonymyeon/backend/consumption/integration/ConsumptionIntegrationTest.java
@@ -1,6 +1,6 @@
 package edonymyeon.backend.consumption.integration;
 
-import static edonymyeon.backend.consumption.integration.steps.ConsumptionSteps.특정_기간의_소비금액을_확인한다;
+import static edonymyeon.backend.support.IntegrationFixture.ConsumptionSteps.특정_기간의_소비금액을_확인한다;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import edonymyeon.backend.consumption.domain.Consumption;
