+++ backend/src/test/java/edonymyeon/backend/support/IntegrationFixture.java
@@ -3,6 +3,8 @@ package edonymyeon.backend.support;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 
+import edonymyeon.backend.member.application.dto.request.PurchaseConfirmRequest;
+import edonymyeon.backend.member.application.dto.request.SavingConfirmRequest;
 import edonymyeon.backend.member.domain.Member;
 import edonymyeon.backend.notification.application.NotificationSender;
 import io.restassured.RestAssured;
@@ -13,6 +15,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.http.MediaType;
 
 @SuppressWarnings("NonAsciiCharacters")
 @IntegrationTest
@@ -90,6 +93,112 @@ public class IntegrationFixture {
 
     private void 알림전송기능을_모킹한다() {
         doNothing().when(notificationSender).sendNotification(any(), any());
+    }
+
+    public class CommentSteps {
+
+        public static ExtractableResponse<Response> 댓글을_생성한다(
+                final Long 게시글_id,
+                final File 이미지,
+                final String 내용,
+                final Member 사용자
+        ) {
+            return RestAssured.given()
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .multiPart("image", 이미지)
+                    .multiPart("comment", 내용)
+                    .when()
+                    .post("/posts/{postId}/comments", 게시글_id)
+                    .then()
+                    .extract();
+        }
+
+        public static ExtractableResponse<Response> 댓글을_삭제한다(
+                final Long 게시글_id,
+                final Long 댓글_id,
+                final Member 사용자
+        ) {
+            return RestAssured.given()
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .when()
+                    .delete("/posts/{postId}/comments/{commentId}", 게시글_id, 댓글_id)
+                    .then()
+                    .extract();
+        }
+
+        public static ExtractableResponse<Response> 게시물에_대한_댓글을_모두_조회한다(
+                final Long 게시글_id,
+                final Member 사용자
+        ) {
+            return RestAssured.given()
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .when()
+                    .get("/posts/{postId}/comments", 게시글_id)
+                    .then()
+                    .extract();
+        }
+    }
+
+    public class ConsumptionSteps {
+
+        public static ExtractableResponse<Response> 특정_기간의_소비금액을_확인한다(final Integer 기간, final Member 사용자) {
+            final ExtractableResponse<Response> 조회_응답 = RestAssured
+                    .given()
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .when()
+                    .get("/consumptions?period-month={periodMonth}", 기간)
+                    .then()
+                    .extract();
+            return 조회_응답;
+        }
+    }
+
+    public class MemberConsumptionSteps {
+
+        public static ExtractableResponse<Response> 구매_확정_요청을_보낸다(
+                final Member 사용자,
+                final Long 게시글_id,
+                final PurchaseConfirmRequest 구매_확정_요청
+        ) {
+            return RestAssured
+                    .given()
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .body(구매_확정_요청)
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .when()
+                    .post("/profile/my-posts/{postId}/purchase-confirm", 게시글_id)
+                    .then()
+                    .extract();
+        }
+
+        public static ExtractableResponse<Response> 절약_확정_요청을_보낸다(
+                final Member 사용자,
+                final Long 게시글_id,
+                final SavingConfirmRequest 절약_확정_요청
+        ) {
+            return RestAssured
+                    .given()
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .body(절약_확정_요청)
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .when()
+                    .post("/profile/my-posts/{postId}/saving-confirm", 게시글_id)
+                    .then()
+                    .extract();
+        }
+
+        public static ExtractableResponse<Response> 확정_취소_요청을_보낸다(final Member 사용자, final Long 게시글_id) {
+            return RestAssured
+                    .given()
+                    .contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .auth().preemptive().basic(사용자.getEmail(), 사용자.getPassword())
+                    .when()
+                    .delete("/profile/my-posts/{postId}/confirm-remove", 게시글_id)
+                    .then()
+                    .log().all()
+                    .extract();
+        }
     }
 }
